+++ StarshipArena/src/PlanetRadar.java
@@ -4,15 +4,15 @@ public class PlanetRadar extends Starship{
 	
 	static Texture tex1 = new Texture("WIP.png");
 	
-	public PlanetRadar(StarshipArena mygame, int spawnx, int spawny){
+	public PlanetRadar(StarshipArena mygame, double spawnx, double spawny){
 		super(mygame, spawnx, spawny);
 	}
 	
-	public PlanetRadar(StarshipArena mygame, int spawnx, int spawny, double spawnangle){
+	public PlanetRadar(StarshipArena mygame, double spawnx, double spawny, double spawnangle){
 		super(mygame, "none", spawnx, spawny, spawnangle);
 	}
 
-	public PlanetRadar(StarshipArena mygame, String newteam, int spawnx, int spawny, double spawnangle){
+	public PlanetRadar(StarshipArena mygame, String newteam, double spawnx, double spawny, double spawnangle){
 		super(mygame, newteam, spawnx, spawny, spawnangle);
 	}
